+++ src/app/index.jsx
@@ -1,6 +1,7 @@
 import { createRoot } from "react-dom/client";
 import { StrictMode } from "react";
 import App from "./App.jsx";
+import { getURLSearchParamSpec } from "./spec_links.js";
 
 const root = createRoot(document.getElementById("app"));
-root.render(<StrictMode><App /></StrictMode>);
+root.render(<StrictMode><App initialSpec={getURLSearchParamSpec()} /></StrictMode>);