+++ src/components/mode_component.jsx
@@ -1,4 +1,4 @@
-export default function({ mode, edit, transform, analyze }) {
+export default function({ mode, edit, transform, analyze, copySpecLink }) {
   function onChange(e) {
     if (e.target.value === "edit") {
       edit();
@@ -9,11 +9,15 @@ export default function({ mode, edit, transform, analyze }) {
 
   return (
     <div id="mode">
-      <input id="mode-edit" type="radio" name="mode" value="edit" checked={mode === "edit"} onChange={onChange} />
-      <label className="left" htmlFor="mode-edit">Edit</label>
-      <input id="mode-transform" type="radio" name="mode" value="transform" checked={mode === "transform"} onChange={onChange} />
-      <label className="right" htmlFor="mode-transform">Transform</label>
+      <div id="mode-switch">
+        <input id="mode-edit" type="radio" name="mode" value="edit" checked={mode === "edit"} onChange={onChange} />
+        <label className="left" htmlFor="mode-edit">Edit</label>
+        <input id="mode-transform" type="radio" name="mode" value="transform" checked={mode === "transform"} onChange={onChange} />
+        <label className="right" htmlFor="mode-transform">Transform</label>
+      </div>
+
       <button id="mode-analyze" disabled={mode !== "edit"} onClick={analyze}>Analyze</button>
+      <button id="mode-copy" onClick={copySpecLink}>Copy Link</button>
     </div>
   );
 }